Moved session persistence to the right package.
